move CycleError to plumbing
